Move language settings to player category
Fixes #1908
