+++ engine/src/main/java/org/terasology/rendering/nui/layers/mainMenu/settings/SettingsMenuScreen.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.rendering.nui.layers.mainMenu;
+package org.terasology.rendering.nui.layers.mainMenu.settings;
 
 import org.terasology.asset.Assets;
 import org.terasology.assets.ResourceUrn;
