+++ engine/src/main/resources/assets/ui/menu/videoMenuScreen.ui
@@ -100,22 +100,6 @@
                                 "id": "vsync"
                             },
                             {
-                                "type": "UILabel",
-                                "text": "${engine:menu#settings-language}: "
-                            },
-                            {
-                                "type": "UIDropdown",
-                                "id": "language"
-                            },
-                            {
-                                "type": "UISpace",
-                                "size": [1, 13]
-                            },
-                            {
-                                "type": "UISpace",
-                                "size": [1, 13]
-                            },
-                            {
                                 "type": "UISpace",
                                 "size": [1, 13]
                             },
